feat: cache 메모리 제한
