+++ cache/redis.conf
@@ -1,0 +1,4 @@
+port 6379
+requirepass dev_password
+maxmemory 256mb
+maxmemory-policy volatile-ttl
